requirements.txt
discord.py
TikTokLive
asyncio